Fix for Primera
Current code ends up invoking Client API with 'tdvv' as False even when 'thin' provisioning is specified by the user. Client API mandates that 'tdvv' and 'compression' must be passed together. Else it results in error. This fix ensures that 'tdvv' is passed to Client API only when specified by the user.